added testcases for csr_preprocessor
